requirements.txt
mido==1.2.10
pygame==2.6.1
matplotlib==3.8.1
networkx==3.1
pyyaml==6.0